+++ app/home/user-management/dashboard/page.tsx
@@ -4,22 +4,18 @@ import DashboardHeader from "@/components/layout/DashboardHeader";
 import {
   Card, CardContent } from "@/components/ui/card";
 import LeadsTable from "@/components/tables/crm/LeadsTable";
+import UsersTable from "@/components/tables/user-management/UsersTable";
 
 
 function UserManagementDashboard() {
-  const [leadsThisMonth, setLeadsThisMonth] = useState(24450);
-  const [leadsLastMonth, setLeadsLastMonth] = useState(200);
-  const [totalLeads, setTotalLeads] = useState(0);
+  const [totalUsers, setTotalUsers] = useState(0);
+  const [totalDepartments, setTotalDepartments] = useState(0);
+  const [deactivatedUsers, setDeactivatedUsers] = useState(0);
 
-  useEffect(() => {
-    setTotalLeads(leadsThisMonth + leadsLastMonth);
-  }, []);
-
-  const leads = [
-    {id: 1, leads: leadsThisMonth, label: "Total Views", hasIncrease: true},
-    {id: 2, leads: "19%", label: "Bounce Rate", hasIncrease: false},
-    {id: 3, leads: totalLeads, label: "Conversion Rate", hasIncrease: false},
-    {id: 4, leads: "30s", label: "Average Engagement Time", hasIncrease: false},
+  const user_stats = [
+    {id: 1, stats: totalUsers, label: "Total Users"},
+    {id: 2, stats: totalDepartments, label: "Total departments"},
+    {id: 3, stats: deactivatedUsers, label: "Deactivated Users"},
   ]
   
   return (
@@ -32,14 +28,13 @@ function UserManagementDashboard() {
         
         <div className="flex flex-col gap-8">
           {/* TOP STATS CARDS */}
-          <div className="mt-8 grid gap-4 md:grid-cols-4">
-            {leads.map((lead) => (
-              <Card key={lead.id} className="flex flex-col justify-end h-[150px]">
+          <div className="mt-8 grid gap-4 md:grid-cols-3">
+            {user_stats.map((user_stat) => (
+              <Card key={user_stat.id} className="flex flex-col justify-end h-[150px]">
                 <CardContent className="flex flex-col items-start">
-                  <div className="text-4xl font-bold">{lead.leads}</div>
+                  <div className="text-4xl font-bold">{user_stat.stats}</div>
                   <div className="flex items-center gap-2">
-                    <p className="text-lg font-semibold text-muted-foreground">{lead.label}</p>
-                    {lead.hasIncrease && <p className="text-xs text-green-600">+12% from last month</p>}
+                    <p className="text-lg font-semibold text-muted-foreground">{user_stat.label}</p>
                   </div>
                 </CardContent>
               </Card>
@@ -47,7 +42,7 @@ function UserManagementDashboard() {
           </div>
 
           {/* LEADS TABLE SECTION */}
-          <LeadsTable table_name="Recent Leads" recentViewOnly={true} />
+          <UsersTable onEdit={() => {}} onDelete={() => {}} onAdd={() => {}} />
         </div>
 
       </div>