+++ lib/auth-client.ts
@@ -7,7 +7,7 @@ export const authClient = createAuthClient({
   baseURL: process.env.NEXT_PUBLIC_APP_URL ?? 'http://localhost:3000',
   plugins: [
     inferAdditionalFields<typeof auth>(),
-    adminClient()
+    adminClient(), 
   ],
 });