+++ next.config.ts
@@ -3,7 +3,18 @@ import type { NextConfig } from "next";
 const nextConfig: NextConfig = {
   /* config options here */
   images: {
-    remotePatterns: [new URL('https://i.imgur.com/**')],
+    remotePatterns: [
+      {
+        protocol: "https",
+        hostname: "www.rland.ph",
+        pathname: "/storage/images/**",
+      },
+      {
+        protocol: "https",
+        hostname: "i.imgur.com",
+        pathname: "/**",
+      },
+    ],
   },
 };