+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.6] - 2026-03-09
+
+### Added
+- User Management: added users management using better-auth
+- Added department ENUM and types
+- Fixed the order of projects shown in the Inventory tab
+- (CMS): Added security tools page 
+
+### Changes
+- Re-arranged the System Setting's layout
+
+### Removed
+- Removed unused tabs/routes in the CMS
+
 ## [0.0.5] - 2026-03-05
 
 ### Added